requirements.txt
Python==3.7.4
Django==3.1.3
djangorestframework==3.12.2
